Finish #valid_moves
This involved having to go back and add a parameter for the board.

The King's threat_map function required getting all capture spots from
other peices that were a threat, and so I needed to be able to pass in
an empty board to get those pieces. It took me a while to figure out
why the king was able to self-check when there was a capture involved.

I also had to add a helper that only showed pawn captures, as the old
\#pawn function primarily showed forward movements with diagonal captures
if available. In a situation where an enemy pawn was on the same rank as
a king, I was getting an option to move the King into a pawn capture square.