Three failing tests from pawn movement.
The problem is that the pawn in question should be pinned by the Queen,
so the pawn can not capture or move forward.

Things to consider:
How do we determine a piece is pinned?
We can check the current piece against an identified threat. Based on
the threatening piece's available movements, that axis of movement
should by the only available axis for the pinned piece. Further, if we
are restricted to one axis, and we cannot capture the threatening piece
on that axis, then we actually cannot move at all as we are not able to
prevent a check.

I also updated the #under_attack? method in Movement to #threaten? as
it more accurately reflects the output of that method. More or less,
is the king on any axis that an enemy piece can move on?