Update move spec
I've added some rationale to why I am approaching this spec file the
way I am.

My concern with relying on Pieces to actually generate the moves in this
test is that the goal for the Move object is to generate a path of cells
as long as they are in bounds. Then inside each piece, I can make determinations
about whether a move conforms to the basic movement rules.

I think this approach is better because, for example, in the case of a
sliding piece, each move will be the whole axis. Because I then have
information about that axis, I can ask questions about whether there's
a friendly obstruction, a capture, etc.

Then I can step through each move and get the capture, or friendly.